got rid of some cruft

+++ resources/migrations/20181007174951-add-images-table.up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE images(
   md5              UUID           PRIMARY KEY,
-  device_id        INTEGER        REFERENCES devices(id),
+  device_id        BIGSERIAL      REFERENCES devices(id),
   path             VARCHAR (355)  NOT NULL,
   created_on       TIMESTAMP      NOT NULL
 );
